chore: move all functions to a individual .hpp and .cpp
