+++ gstreamerTypes.hpp
@@ -108,176 +108,6 @@ namespace gstreamer {
          */
         bool binary_in = false;
     };
-
-    /** Response from each connected peer to the RTP pipeline */
-
-    struct RTPMonitoredSession {
-        std::string name;
-        int session_id;
-    };
-
-    struct RTPMonitoredSessions {
-        std::string rtpbin_name;
-        std::vector<RTPMonitoredSession> sessions;
-    };
-
-    struct RTPPeerReceiverReport {
-        /** The SSRC of this source */
-        uint16_t ssrc = 0;
-        /** The SSRC of the sender of this Receiver Report */
-        uint16_t sender_ssrc = 0;
-        /** Lost 8-bit fraction */
-        float fractionlost = 0;
-        /** Lost packets */
-        int32_t packetslost = 0;
-        /** Highest received seqnum */
-        uint32_t exthighestseq = 0;
-        /** Reception jitter if clock rate is present */
-        base::Time jitter;
-        /** Last Sender Report time */
-        base::Time lsr;
-        /** Delay since last Sender Report */
-        base::Time dlsr;
-        /** The round-trip time */
-        base::Time round_trip;
-    };
-
-    /** The RTP Session Statistics
-     *
-     * This is the response when the stats is present in the session
-     */
-    struct RTPSourceStatistics {
-        /** The label of the stream */
-        std::string stream_name;
-
-        // The following fields are always present.
-        /** SSRC of this source */
-        uint16_t ssrc = 0;
-
-        enum Flags {
-            SOURCE_IS_FROM_SELF = 0x1,
-            SOURCE_IS_VALIDATED = 0x2,
-            BYE_RECEIVED = 0x4,
-            SOURCE_IS_CSRC = 0x8
-        };
-
-        /** @meta bitfield /gstreamer/RTPSourceStatistics/Flags */
-        uint8_t flags = 0;
-
-        /** Returns a string corresponding to the \see flags.
-         * The method is located inside RTPTask.cpp */
-        std::string flagsToString();
-
-        /** RTP Bitfield from the received GstStructure confirmations as a string */
-        std::string confirmations = "";
-
-        /** First seqnum if known */
-        int32_t seqnum_base = 0;
-        /** Clock rate of the media */
-        int32_t clock_rate = 0;
-    };
-
-    struct RTPSenderStatistics {
-        RTPSourceStatistics source_stats;
-
-        // The following fields make sense for internal sources
-        /** Number of payload bytes we sent */
-        uint64_t payload_bytes_sent = 0;
-        /** Number of packets we sent */
-        uint64_t packets_sent = 0;
-
-        // The following fields make sense for non-internal sources and will only increase
-        // when "is-sender" is TRUE.
-        /** Total number of payload bytes received */
-        uint64_t payload_bytes_received = 0;
-        /** Total number of packets received */
-        uint64_t packets_received = 0;
-        /** Total number of bytes received including lower level headers overhead */
-        uint64_t bytes_received = 0;
-
-        // Following fields are updated when "is-sender" is TRUE.
-        /** Bitrate in bits/sec */
-        uint64_t bitrate = 0;
-        /** Estimated amount of packets lost */
-        int32_t packets_lost = 0;
-        /** Estimated Jitter if clock rate is available */
-        base::Time jitter;
-
-        // The last Sender Report report this source sent. Only updates when "is-sender"
-        // is TRUE.
-        /** The source has sent Sender Report */
-        bool have_sr = false;
-        /** NTP Time of the Sender Report */
-        base::Time sr_ntptime;
-        /** RTP Time of Sender Report in clock rate units*/
-        uint32_t sr_rtptime_in_clock_rate_units;
-        /** The number of bytes in the Sender Report */
-        uint32_t sr_octet_count = 0;
-        /** The number of packets in the Sender Report */
-        uint32_t sr_packet_count = 0;
-
-        // The following fields are only present for non-internal sources and represent
-        // the content of the last Report Block Buffer packet that was sent to this source. These
-        // values are only updated when the source is sending.
-        /** Confirmation if a Report Block Buffer has been sent */
-        bool sent_receiver_block = false;
-        /** Lost Packets */
-        uint32_t sent_receiver_block_packetslost = 0;
-        /** calculated lost 8-bit fraction */
-        float sent_receiver_block_fractionlost = 0;
-        /** Last seen seqnum */
-        uint32_t sent_receiver_block_exthighestseq = 0;
-        /** Jitter if clock rate is available*/
-        base::Time sent_receiver_block_jitter;
-        /** Last Sender Report time */
-        base::Time sent_receiver_block_lsr;
-        /** delay since last Sender Report */
-        base::Time sent_receiver_block_dlsr;
-
-        // Documentation not available
-        uint32_t sent_picture_loss_count = 0;
-        uint32_t sent_full_image_request_count = 0;
-        uint32_t sent_nack_count = 0;
-        uint32_t received_full_image_request_count = 0;
-
-        // The following field is present only for internal sources and contains an array
-        // of the the most recent receiver reports from each peer. In unicast scenarios
-        // this will be a single entry that is identical to the data provided by the
-        // have-rb and rb-* fields, but in multicast there will be one entry in the array
-        // for each peer that is sending receiver reports.
-        std::vector<RTPPeerReceiverReport> peer_receiver_reports;
-    };
-
-    struct RTPReceiverStatistics {
-        RTPSourceStatistics source_stats;
-
-        // The following fields are only present when known.
-        /** Origin of last received RTP Packet - Only present when known */
-        std::string rtp_from = std::string("");
-        /** Origin of last received RTP Packet - Only present when known */
-        std::string rtcp_from = std::string("");
-
-        // Documentation not available
-        uint32_t received_picture_loss_count = 0;
-        uint32_t received_nack_count = 0;
-        uint64_t received_packet_rate = 0;
-    };
-
-    struct RTPSessionStatistics {
-        uint64_t rtx_drop_count = 0;
-        uint64_t sent_nack_count = 0;
-        uint64_t recv_nack_count = 0;
-        uint64_t sent_rtx_req_count = 0;
-        uint64_t recv_rtx_req_count = 0;
-
-        std::vector<RTPSenderStatistics> sender_stats;
-        std::vector<RTPReceiverStatistics> receiver_stats;
-    };
-
-    struct RTPStatistics {
-        base::Time time;
-        std::vector<RTPSessionStatistics> statistics;
-    };
 }
 
 #endif
