+++ tasks/RTPTask.hpp
@@ -50,149 +50,12 @@ namespace gstreamer {
          */
         RTPSessionStatistics extractRTPSessionStats(GstElement* session);
 
-        /** Extract the stats which are always present on a RTP Source
-         *
-         * @param gst_stats [GstStructure*] stats from a given source
-         * @param rtpbin_name [std::string] from a given source
-         * @return [RTPSourceStatistics] RTP Source obligatory stats
-         */
-        static RTPSourceStatistics extractRTPSourceStats(GstStructure const* gst_stats, std::string rtpbin_name);
-
-        /** Extract the Sender Stats from a RTP Source.
-         *
-         *  The sender stats are all the parameters that are updated when is-sender
-         *  is TRUE and were defined by checking the documentation from
-         *  https://gstreamer.freedesktop.org/documentation/rtpmanager/RTPSource.html?gi-language=c#RTPSource:stats
-         *
-         * @param stats [GstStructure*] stats from a given source
-         * @param clock_rate [uint32_t] in clock rate units from RTPSource
-         * @return [RTPSenderStatistics] RTP Sender Stats
-         */
-        static RTPSenderStatistics extractRTPSenderStats(GstStructure const* stats,
-            uint32_t clock_rate);
-
-        /** Extract the Stats from a RTP Receiver
-         *
-         *  The receiver stats are all the parameters that are updated when is-sender
-         *  is FALSE and were defined by checking the documentation from
-         *  https://gstreamer.freedesktop.org/documentation/rtpmanager/RTPSource.html?gi-language=c#RTPSource:stats
-         *
-         * @param stats [GstStructure*] stats from a given source
-         * @return [RTPSenderStatistics] RTP Receiver Stats
-         */
-        static RTPReceiverStatistics extractRTPReceiverStats(GstStructure const* stats);
-
-        /** Fetch the gbooleans and converts it to bool
-         *
-         *  @param structure [GstStructure*]
-         *  @param fieldname [char*]
-         *  @return [bool]
-         */
-        static bool fetchBoolean(const GstStructure* structure, const char* fieldname);
-
-        /** Fetchs a 64 unsigned int
-         *
-         *  @param structure[GstStructure*]
-         *  @param fieldname [char*] The name of the field containing the variable
-         *  @return [uint64_t]
-         */
-        static uint64_t fetch64UnsignedInt(const GstStructure* structure,
-            const char* fieldname);
-
-        /** Fetchs a 32 unsigned int
-         *
-         *  @param structure[GstStructure*]
-         *  @param fieldname [char*] The name of the field containing the variable
-         *  @return [uint32_t]
-         */
-        static uint32_t fetchUnsignedInt(const GstStructure* structure,
-            const char* fieldname);
-
-        /** Fetchs a int
-         *
-         *  @param structure[GstStructure*]
-         *  @param fieldname [char*] The name of the field containing the variable
-         *  @return [int32_t]
-         */
-        static int32_t fetchInt(const GstStructure* structure, const char* fieldname);
-
-        /** Fetchs a gstring and converts it to std::string
-         *
-         *  @param structure [GstStructure*]
-         *  @param fieldname [char*] The name of the field containing the variable
-         *  @return [std::string]
-         */
-        static std::string fetchString(const GstStructure* structure,
-            const char* fieldname);
-
-        /** Updates flags of a RTPSourceStatistics struct
-         *
-         *  @param gst_stats [GstStructure]
-         *  @return [uint8_t]
-         */
-        static uint8_t fetchFlags(const GstStructure* gst_stats);
-
-        /** NTP Timestamp format to base::Time from Unix Epoch
-         *
-         * @param ntp_timestamp [uint64_t] NTP Timestamp 32.32
-         * @return [base::Time] Time from Unix epoch
-         */
-        static base::Time ntpToUnixEpoch(uint64_t ntp_timestamp);
-
-        /**
-         * Transforms a Last Sender Report Time (NTP Format 16.16 Fixed Point) to
-         * microseconds from UnixEpoch using the NTP Timestamp from the report to
-         * define the fixed point (16 most significative bits).
-         *
-         * This works because the fixed time is the middle 32 bits out of 64
-         * in the NTP timestamp.
-         *
-         * For further explanations on how this calculation works refer to
-         * https://www.ietf.org/rfc/rfc1889.txt
-         * pages 23-24(Diagram and NTP Timestamp) and 27 (Last SR).
-         *
-         * @param ntp_timestamp NTP Timestamp from the RTCP Packet
-         * @param ntp_short Last SR Time (NTP Format 16.16 Fixed Point)
-         * @return [base::Time] Time from Unix epoch
-         */
-        static base::Time lsrTimeToUnixEpoch(uint64_t ntp_timestamp, uint32_t ntp_short);
-
-        /** NTP Short Format to base::Time
-         *
-         * @param ntp_short NTP Short Format 16.16 Fixed Point
-         * @return [base::Time] Time from Unix epoch
-         */
-        static base::Time NTPShortToTime(uint32_t ntp_short);
-
-        /** Converts Jitter from Clock Rate Units to base::Time
-         *
-         * @param jitter Jitter in Clock Rate Units
-         * @param clock_rate Units from RTP Source
-         * @return [base::Time] Time from Unix epoch
-         */
-        static base::Time jitterToTime(uint32_t jitter, uint32_t clock_rate);
-
-        /** Extracts a array of receiver reports contained on a GST_TYPE_LIST of
-         * GstStructures on the "received-rr" field of a RTPSource
-         *
-         * @param structure [GstStructure*]
-         * @param clock_rate [uint32_t] Units
-         * @param ntp_timestamp NTP Timestamp of the RTP Source 32.32 format
-         * @return [std::vector<RTPPeerReceiverReport>] Received peer reports
-         * */
-        static std::vector<RTPPeerReceiverReport> extractPeerReceiverReports(
-            const GstStructure* structure, uint32_t clock_rate, uint64_t ntp_timestamp);
-
         /** RTP Monitored Settings */
         RTPMonitoredSessions m_rtp_monitored_sessions;
         /** RTP bin element */
         GstElement* m_bin;
         /** RTP sessions element */
         std::vector<GstElement*> m_rtp_sessions;
-        /** Source stats which are always present */
-        RTPSourceStatistics m_source_stats;
-        /** NTP Timestamp */
-        uint64_t m_ntp_timestamp;
 
         /**
          * Hook called when the state machine transitions from PreOperational to
