+++ test/CMakeLists.txt
@@ -1,7 +1,8 @@
 rock_gtest(test_suite suite.cpp
-                      test_RTPTask.cpp
+            test_RTPTask.cpp
+    SOURCES ../tasks/RTPHelpers.cpp
+    HEADERS ../tasks/RTPHelpers.hpp
     DEPS_PKGCONFIG gstreamer-1.0
-                   gstreamer-tasks-gnulinux
                    base-types
 )
